Refactor server into package modules
