+++ bandtrack/api/__init__.py
@@ -66,36 +66,15 @@ import os
 import sqlite3
 try:
     import psycopg2  # type: ignore
-    import psycopg2.extras as psycopg2_extras  # type: ignore
-    Psycopg2Error = psycopg2.Error  # type: ignore[attr-defined]
 except ModuleNotFoundError:  # pragma: no cover - optional dependency
     psycopg2 = None  # type: ignore
-    psycopg2_extras = None  # type: ignore
-    class Psycopg2Error(Exception):
-        pass
-from contextlib import contextmanager
-import secrets
-import string
-import hashlib
-import hmac
-import time
 import datetime
 import gzip
 import urllib.parse
 import mimetypes
 import io
-from email.parser import BytesParser
-from email.policy import default
-import subprocess
-import shlex
 import asyncio
 import threading
-import sqlparse
-from sqlparse import tokens as T
-try:
-    import websockets  # type: ignore
-except ModuleNotFoundError:  # pragma: no cover - optional dependency
-    websockets = None  # type: ignore
 from http import HTTPStatus
 from http.server import ThreadingHTTPServer, BaseHTTPRequestHandler
 try:
@@ -104,18 +83,35 @@ try:
 except ImportError:  # pragma: no cover - optional dependency
     canvas = None
     A4 = None
-from scripts.migrate_to_multigroup import migrate as migrate_to_multigroup
-from scripts.migrate_suggestion_votes import migrate as migrate_suggestion_votes
-from scripts.migrate_performance_location import migrate as migrate_performance_location
-from scripts.migrate_sessions_group_id import migrate as migrate_sessions_group_id
+try:
+    import websockets  # type: ignore
+except ModuleNotFoundError:  # pragma: no cover - optional dependency
+    websockets = None  # type: ignore
 
-#############################
-# Database initialisation
-#############################
+from bandtrack.db import (
+    get_db_connection,
+    open_db_connection,
+    execute_write,
+    safe_commit,
+    PartitionDAO,
+    get_partition_dao,
+    init_db,
+    _using_postgres,
+)
+from bandtrack.auth import (
+    hash_password,
+    verify_password,
+    generate_invitation_code,
+    generate_unique_invitation_code,
+    generate_session,
+    get_user_by_session,
+    delete_session,
+    log_event,
+)
+from bandtrack.utils import sanitize_name, scan_for_viruses, parse_multipart_form_data
 
-DB_FILENAME = os.path.join(os.path.dirname(__file__), 'bandtrack.db')
-
-UPLOADS_ROOT = os.path.join(os.path.dirname(__file__), 'uploads', 'partitions')
+# Path for uploaded partition files
+UPLOADS_ROOT = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'uploads', 'partitions')
 
 # Maximum allowed size for uploaded partition files (default 5 MB)
 MAX_PARTITION_SIZE = int(os.environ.get('MAX_PARTITION_SIZE', 5 * 1024 * 1024))
@@ -158,710 +154,6 @@ def start_ws_server(host: str, port: int) -> None:
     server = websockets.serve(ws_handler, host, port)
     WS_LOOP.run_until_complete(server)
     WS_LOOP.run_forever()
-
-
-def _using_postgres() -> bool:
-    """Return True if a PostgreSQL DSN is configured."""
-    return bool(
-        os.environ.get("DATABASE_URL")
-        or os.environ.get("DB_HOST")
-    )
-
-
-def _pg_dsn() -> str:
-    """Build a PostgreSQL connection string from environment variables."""
-    url = os.environ.get("DATABASE_URL")
-    if url:
-        return url
-    host = os.environ.get("DB_HOST", "localhost")
-    port = os.environ.get("DB_PORT", "5432")
-    user = os.environ.get("DB_USER", "postgres")
-    password = os.environ.get("DB_PASSWORD", "")
-    dbname = os.environ.get("DB_NAME", "postgres")
-    return f"postgresql://{user}:{password}@{host}:{port}/{dbname}"
-
-
-def to_psycopg2_params(query: str, params: tuple) -> tuple[str, tuple]:
-    """Convert SQLite-style '?' placeholders to psycopg2 parameters.
-
-    Uses ``sqlparse`` to avoid replacing question marks that appear inside
-    string literals or comments. Raises ``ValueError`` if the number of
-    placeholders does not match ``params``.
-    """
-    parsed = sqlparse.parse(query)
-    if not parsed:
-        return query, params
-    tokens: list[str] = []
-    count = 0
-    for token in parsed[0].flatten():
-        if token.ttype == T.Name.Placeholder and token.value == "?":
-            tokens.append("%s")
-            count += 1
-        else:
-            tokens.append(token.value)
-    if count != len(params):
-        raise ValueError("Mismatched number of parameters")
-    return "".join(tokens), params
-
-
-def execute_write(target, sql, params=()):
-    """Execute a SQL statement with retry on database locks."""
-    if _using_postgres():
-        sql, params = to_psycopg2_params(sql, params)
-        try:
-            result = target.execute(sql, params)
-        except Psycopg2Error:
-            target.connection.rollback()
-            raise
-        if sql.lstrip().upper().startswith("INSERT") and not getattr(target, "lastrowid", None):
-            target.execute("SELECT LASTVAL()"); target.lastrowid = target.fetchone()[0]
-        return result
-    delay = 0.05
-    for attempt in range(5):
-        try:
-            return target.execute(sql, params)
-        except sqlite3.OperationalError as e:
-            if "database is locked" in str(e).lower() and attempt < 4:
-                time.sleep(delay)
-                delay *= 2
-                continue
-            raise
-
-
-def safe_commit(conn):
-    """Commit the current transaction, rolling back on failure."""
-    if _using_postgres():
-        try:
-            conn.commit()
-        except Psycopg2Error:
-            conn.rollback()
-            raise
-    else:
-        conn.commit()
-
-
-@contextmanager
-def get_db_connection():
-    """Yield a database connection with foreign keys enabled."""
-    if _using_postgres():
-        if psycopg2 is None:
-            raise RuntimeError("PostgreSQL support requires installing psycopg2")
-        conn = psycopg2.connect(_pg_dsn(), cursor_factory=psycopg2_extras.RealDictCursor)
-    else:
-        conn = sqlite3.connect(DB_FILENAME, check_same_thread=False, timeout=30)
-        execute_write(conn, 'PRAGMA foreign_keys = ON')
-        execute_write(conn, 'PRAGMA journal_mode=WAL')
-        execute_write(conn, 'PRAGMA busy_timeout=30000')
-        conn.row_factory = sqlite3.Row
-    try:
-        yield conn
-        safe_commit(conn)
-    finally:
-        conn.close()
-
-
-def open_db_connection():
-    """Return a new database connection with the standard settings applied."""
-    if _using_postgres():
-        if psycopg2 is None:
-            raise RuntimeError("PostgreSQL support requires installing psycopg2")
-        return psycopg2.connect(_pg_dsn(), cursor_factory=psycopg2_extras.RealDictCursor)
-    conn = sqlite3.connect(DB_FILENAME, check_same_thread=False, timeout=30)
-    execute_write(conn, 'PRAGMA foreign_keys = ON')
-    execute_write(conn, 'PRAGMA journal_mode=WAL')
-    execute_write(conn, 'PRAGMA busy_timeout=30000')
-    conn.row_factory = sqlite3.Row
-    return conn
-
-
-class PartitionDAO:
-    """Data access helper for the partitions table."""
-
-    def __init__(self, conn):
-        self.conn = conn
-
-    def create(self, rehearsal_id: int, path: str, display_name: str, uploader_id: int) -> int:
-        cur = self.conn.cursor()
-        execute_write(cur, 
-            'INSERT INTO partitions (rehearsal_id, path, display_name, uploader_id) VALUES (?, ?, ?, ?)',
-            (rehearsal_id, path, display_name, uploader_id),
-        )
-        safe_commit(self.conn)
-        return cur.lastrowid
-
-    def list_by_rehearsal(self, rehearsal_id: int) -> list:
-        cur = self.conn.cursor()
-        execute_write(cur, 'SELECT * FROM partitions WHERE rehearsal_id = ?', (rehearsal_id,))
-        return cur.fetchall()
-
-    def delete(self, partition_id: int) -> None:
-        cur = self.conn.cursor()
-        execute_write(cur, 'DELETE FROM partitions WHERE id = ?', (partition_id,))
-        safe_commit(self.conn)
-
-
-def get_partition_dao(conn: sqlite3.Connection) -> PartitionDAO:
-    return PartitionDAO(conn)
-
-def init_db():
-    """Create tables if they do not already exist and insert the
-    default settings row.  This function is idempotent."""
-    new_db = (not _using_postgres()) and not os.path.exists(DB_FILENAME)
-    with get_db_connection() as conn:
-        cur = conn.cursor()
-
-        def run(stmt: str) -> None:
-            if _using_postgres():
-                stmt = (
-                    stmt.replace("INTEGER PRIMARY KEY AUTOINCREMENT", "SERIAL PRIMARY KEY")
-                    .replace("BLOB", "BYTEA")
-                    .replace("DATETIME", "TIMESTAMP")
-                )
-            execute_write(cur, stmt)
-
-        # Groups allow multiple band configurations and are owned by a user
-        groups_stmt = (
-            '''CREATE TABLE IF NOT EXISTS groups (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   name TEXT NOT NULL,
-                   invitation_code TEXT NOT NULL UNIQUE,
-                   description TEXT,
-                   logo_url TEXT,
-                   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   owner_id INTEGER NOT NULL,
-                   FOREIGN KEY (owner_id) REFERENCES users(id)
-               );'''
-        )
-        if _using_postgres():
-            groups_stmt = groups_stmt.replace(
-                ',\n                   FOREIGN KEY (owner_id) REFERENCES users(id)',
-                '',
-            )
-        run(groups_stmt)
-
-        # Users table: store username, salt and password hash
-        run(
-            '''CREATE TABLE IF NOT EXISTS users (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   username TEXT NOT NULL UNIQUE,
-                   salt BLOB NOT NULL,
-                   password_hash BLOB NOT NULL,
-                   role TEXT NOT NULL DEFAULT 'user',
-                   last_group_id INTEGER,
-                   notify_uploads INTEGER NOT NULL DEFAULT 1,
-                   FOREIGN KEY (last_group_id) REFERENCES groups(id)
-               );'''
-        )
-        if _using_postgres():
-            try:
-                run(
-                    'ALTER TABLE groups ADD CONSTRAINT fk_groups_owner '
-                    'FOREIGN KEY (owner_id) REFERENCES users(id)'
-                )
-            except Exception:
-                pass
-        # WebAuthn credentials associated with users
-        run(
-            '''CREATE TABLE IF NOT EXISTS users_webauthn (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   user_id INTEGER NOT NULL,
-                   credential_id TEXT NOT NULL UNIQUE,
-                   FOREIGN KEY (user_id) REFERENCES users(id)
-               );'''
-        )
-        # Suggestions: simple list of suggestions with optional URL and creator
-        run(
-            '''CREATE TABLE IF NOT EXISTS suggestions (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   title TEXT NOT NULL,
-                   author TEXT,
-                   youtube TEXT,
-                   url TEXT,
-                   version_of TEXT,
-                   likes INTEGER NOT NULL DEFAULT 0,
-                   creator_id INTEGER NOT NULL,
-                   group_id INTEGER NOT NULL,
-                   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   FOREIGN KEY (creator_id) REFERENCES users(id),
-                   FOREIGN KEY (group_id) REFERENCES groups(id)
-               );'''
-        )
-        # Individual suggestion votes per user
-        run(
-            '''CREATE TABLE IF NOT EXISTS suggestion_votes (
-                   suggestion_id INTEGER NOT NULL,
-                   user_id INTEGER NOT NULL,
-                   PRIMARY KEY (suggestion_id, user_id),
-                   FOREIGN KEY (suggestion_id) REFERENCES suggestions(id),
-                   FOREIGN KEY (user_id) REFERENCES users(id)
-               );'''
-        )
-        # Rehearsals: store levels and notes per user as JSON strings.  Include
-        # optional author, YouTube/Spotify links and audio notes JSON.  The
-        # ``audio_notes_json`` field stores base64‑encoded audio notes per user.
-        run(
-            '''CREATE TABLE IF NOT EXISTS rehearsals (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   title TEXT NOT NULL,
-                   author TEXT,
-                   youtube TEXT,
-                   spotify TEXT,
-                   version_of TEXT,
-                   levels_json TEXT DEFAULT '{}',
-                   notes_json TEXT DEFAULT '{}',
-                   audio_notes_json TEXT DEFAULT '{}',
-                   mastered INTEGER NOT NULL DEFAULT 0,
-                   creator_id INTEGER NOT NULL,
-                   group_id INTEGER NOT NULL,
-                   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   FOREIGN KEY (creator_id) REFERENCES users(id),
-                   FOREIGN KEY (group_id) REFERENCES groups(id)
-               );'''
-        )
-        # Partitions associated with rehearsals
-        run(
-            '''CREATE TABLE IF NOT EXISTS partitions (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   rehearsal_id INTEGER NOT NULL,
-                   path TEXT NOT NULL,
-                   display_name TEXT NOT NULL,
-                   uploader_id INTEGER NOT NULL,
-                   uploaded_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   FOREIGN KEY (rehearsal_id) REFERENCES rehearsals(id) ON DELETE CASCADE,
-                   FOREIGN KEY (uploader_id) REFERENCES users(id)
-               );'''
-        )
-        # Performances: contains name, date and a JSON array of rehearsal IDs
-        run(
-            '''CREATE TABLE IF NOT EXISTS performances (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   name TEXT NOT NULL,
-                   date TEXT NOT NULL,
-                   location TEXT,
-                   songs_json TEXT DEFAULT '[]',
-                   creator_id INTEGER NOT NULL,
-                   group_id INTEGER NOT NULL,
-                   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   FOREIGN KEY (creator_id) REFERENCES users(id),
-                   FOREIGN KEY (group_id) REFERENCES groups(id)
-               );'''
-        )
-        # Rehearsal events: standalone rehearsal occurrences with date and
-        # location information.  These are distinct from the "rehearsals"
-        # table above, which stores songs to rehearse.
-        run(
-            '''CREATE TABLE IF NOT EXISTS rehearsal_events (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   date TEXT NOT NULL,
-                   location TEXT,
-                   group_id INTEGER NOT NULL,
-                   creator_id INTEGER NOT NULL,
-                   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   FOREIGN KEY (group_id) REFERENCES groups(id),
-                   FOREIGN KEY (creator_id) REFERENCES users(id)
-               );'''
-        )
-        # Memberships link users to groups with a role and optional nickname
-        run(
-            '''CREATE TABLE IF NOT EXISTS memberships (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   user_id INTEGER NOT NULL,
-                   group_id INTEGER NOT NULL,
-                   role TEXT NOT NULL,
-                   nickname TEXT,
-                   joined_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   active INTEGER NOT NULL DEFAULT 1,
-                   FOREIGN KEY (user_id) REFERENCES users(id),
-                   FOREIGN KEY (group_id) REFERENCES groups(id)
-               );'''
-        )
-        # Index for quick lookup of a membership by user and group
-        run('CREATE UNIQUE INDEX IF NOT EXISTS idx_memberships_user_group ON memberships(user_id, group_id)')
-        # Settings: single row with group name, dark mode flag and optional
-        # next rehearsal info.  "template" selects the UI theme.
-        run(
-            '''CREATE TABLE IF NOT EXISTS settings (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   group_id INTEGER NOT NULL UNIQUE,
-                   group_name TEXT NOT NULL,
-                   dark_mode INTEGER NOT NULL DEFAULT 1,
-                   template TEXT NOT NULL DEFAULT 'classic',
-                   FOREIGN KEY (group_id) REFERENCES groups(id)
-               );'''
-        )
-        # Sessions: store session token, associated user and expiry timestamp (epoch)
-        sessions_stmt = (
-            '''CREATE TABLE IF NOT EXISTS sessions (
-                   token TEXT PRIMARY KEY,
-                   user_id INTEGER NOT NULL,
-                   expires_at INTEGER NOT NULL,
-                   {group_id}
-                   FOREIGN KEY (user_id) REFERENCES users(id)
-               );'''
-        )
-        if _using_postgres():
-            sessions_stmt = sessions_stmt.format(
-                group_id='group_id INTEGER,\n                   '
-            )
-        else:
-            sessions_stmt = sessions_stmt.format(group_id='')
-        run(sessions_stmt)
-        # Logs: record key user actions
-        run(
-            '''CREATE TABLE IF NOT EXISTS logs (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   timestamp DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   user_id INTEGER,
-                   action TEXT NOT NULL,
-                   metadata TEXT,
-                   FOREIGN KEY (user_id) REFERENCES users(id)
-               );'''
-        )
-        # Notifications: store messages for users
-        run(
-            '''CREATE TABLE IF NOT EXISTS notifications (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   user_id INTEGER NOT NULL,
-                   message TEXT NOT NULL,
-                   created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-                   FOREIGN KEY (user_id) REFERENCES users(id)
-               );'''
-        )
-        # Push subscriptions: store Web Push subscriptions per user
-        run(
-            '''CREATE TABLE IF NOT EXISTS push_subscriptions (
-                   id INTEGER PRIMARY KEY AUTOINCREMENT,
-                   user_id INTEGER NOT NULL,
-                   endpoint TEXT NOT NULL UNIQUE,
-                   p256dh TEXT NOT NULL,
-                   auth TEXT NOT NULL,
-                   FOREIGN KEY (user_id) REFERENCES users(id)
-               );'''
-        )
-        safe_commit(conn)
-
-    if _using_postgres():
-        return
-
-    # Ensure additional columns are present in existing databases.  SQLite
-    # will raise an OperationalError if a column already exists; we
-    # silently ignore such errors.  Users table: role, last_group_id and
-    # notify_uploads preference
-    conn = None
-    try:
-        conn = open_db_connection()
-        cur = conn.cursor()
-        execute_write(cur, 'PRAGMA table_info(users)')
-        columns = [row['name'] for row in cur.fetchall()]
-        if 'role' not in columns:
-            execute_write(cur, "ALTER TABLE users ADD COLUMN role TEXT NOT NULL DEFAULT 'user'")
-        if 'last_group_id' not in columns:
-            execute_write(cur, 'ALTER TABLE users ADD COLUMN last_group_id INTEGER')
-        if 'notify_uploads' not in columns:
-            execute_write(cur, 'ALTER TABLE users ADD COLUMN notify_uploads INTEGER NOT NULL DEFAULT 1')
-        safe_commit(conn)
-    except Exception:
-        pass
-    finally:
-        if conn is not None:
-            conn.close()
-    # Suggestions table: ensure additional columns exist
-    conn = None
-    try:
-        conn = open_db_connection()
-        cur = conn.cursor()
-        execute_write(cur, 'PRAGMA table_info(suggestions)')
-        s_columns = [row['name'] for row in cur.fetchall()]
-        if 'author' not in s_columns:
-            execute_write(cur, 'ALTER TABLE suggestions ADD COLUMN author TEXT')
-        if 'youtube' not in s_columns:
-            execute_write(cur, 'ALTER TABLE suggestions ADD COLUMN youtube TEXT')
-        if 'version_of' not in s_columns:
-            execute_write(cur, 'ALTER TABLE suggestions ADD COLUMN version_of TEXT')
-        if 'likes' not in s_columns:
-            execute_write(cur, 'ALTER TABLE suggestions ADD COLUMN likes INTEGER NOT NULL DEFAULT 0')
-        if 'group_id' not in s_columns:
-            execute_write(cur, 'ALTER TABLE suggestions ADD COLUMN group_id INTEGER')
-            # Populate group_id using creator's active group
-            execute_write(cur, 
-                '''UPDATE suggestions SET group_id = (
-                        SELECT group_id FROM memberships m
-                        WHERE m.user_id = suggestions.creator_id AND m.active = 1
-                        ORDER BY m.group_id LIMIT 1
-                    ) WHERE group_id IS NULL'''
-            )
-        safe_commit(conn)
-    except Exception:
-        pass
-    finally:
-        if conn is not None:
-            conn.close()
-    # Rehearsals table: ensure additional columns exist
-    conn = None
-    try:
-        conn = open_db_connection()
-        cur = conn.cursor()
-        execute_write(cur, 'PRAGMA table_info(rehearsals)')
-        r_columns = [row['name'] for row in cur.fetchall()]
-        if 'author' not in r_columns:
-            execute_write(cur, 'ALTER TABLE rehearsals ADD COLUMN author TEXT')
-        if 'audio_notes_json' not in r_columns:
-            execute_write(cur, "ALTER TABLE rehearsals ADD COLUMN audio_notes_json TEXT DEFAULT '{}'")
-        if 'sheet_music_json' in r_columns and not new_db:
-            try:
-                execute_write(cur, 'ALTER TABLE rehearsals DROP COLUMN sheet_music_json')
-            except sqlite3.OperationalError:
-                pass
-        if 'mastered' not in r_columns:
-            execute_write(cur, 'ALTER TABLE rehearsals ADD COLUMN mastered INTEGER NOT NULL DEFAULT 0')
-        if 'version_of' not in r_columns:
-            execute_write(cur, 'ALTER TABLE rehearsals ADD COLUMN version_of TEXT')
-        if 'group_id' not in r_columns:
-            execute_write(cur, 'ALTER TABLE rehearsals ADD COLUMN group_id INTEGER')
-            execute_write(cur, 
-                '''UPDATE rehearsals SET group_id = (
-                        SELECT group_id FROM memberships m
-                        WHERE m.user_id = rehearsals.creator_id AND m.active = 1
-                        ORDER BY m.group_id LIMIT 1
-                    ) WHERE group_id IS NULL'''
-            )
-        safe_commit(conn)
-    except Exception:
-        pass
-    finally:
-        if conn is not None:
-            conn.close()
-
-    # Performances table: ensure group_id column exists
-    conn = None
-    try:
-        conn = open_db_connection()
-        cur = conn.cursor()
-        execute_write(cur, 'PRAGMA table_info(performances)')
-        p_columns = [row['name'] for row in cur.fetchall()]
-        if 'group_id' not in p_columns:
-            execute_write(cur, 'ALTER TABLE performances ADD COLUMN group_id INTEGER')
-            execute_write(cur, 
-                '''UPDATE performances SET group_id = (
-                        SELECT group_id FROM memberships m
-                        WHERE m.user_id = performances.creator_id AND m.active = 1
-                        ORDER BY m.group_id LIMIT 1
-                    ) WHERE group_id IS NULL'''
-            )
-        safe_commit(conn)
-    except Exception:
-        pass
-    finally:
-        if conn is not None:
-            conn.close()
-
-    # Settings table: ensure newer columns exist. Older databases may lack
-    # the "template" field, so add it if missing.
-    conn = None
-    try:
-        conn = open_db_connection()
-        cur = conn.cursor()
-        execute_write(cur, 'PRAGMA table_info(settings)')
-        settings_columns = [row['name'] for row in cur.fetchall()]
-        if 'template' not in settings_columns:
-            execute_write(cur, "ALTER TABLE settings ADD COLUMN template TEXT NOT NULL DEFAULT 'classic'")
-        if 'group_id' not in settings_columns:
-            execute_write(cur, 'ALTER TABLE settings ADD COLUMN group_id INTEGER')
-            execute_write(cur, 'UPDATE settings SET group_id = 1 WHERE group_id IS NULL')
-        safe_commit(conn)
-    except Exception:
-        pass
-    finally:
-        if conn is not None:
-            conn.close()
-
-    # Sessions table: ensure a group_id column exists to store the active
-    # group for a session.
-    conn = None
-    try:
-        conn = open_db_connection()
-        cur = conn.cursor()
-        execute_write(cur, 'PRAGMA table_info(sessions)')
-        sess_columns = [row['name'] for row in cur.fetchall()]
-        if 'group_id' not in sess_columns:
-            execute_write(cur, 'ALTER TABLE sessions ADD COLUMN group_id INTEGER')
-        safe_commit(conn)
-    except Exception:
-        pass
-    finally:
-        if conn is not None:
-            conn.close()
-
-#############################
-# Helper functions
-#############################
-
-def hash_password(password: str, salt: bytes | None = None) -> tuple[bytes, bytes]:
-    """Hash a password with PBKDF2.  If ``salt`` is None, a new 16‑byte salt
-    is generated.  Returns a tuple of (salt, password_hash)."""
-    if salt is None:
-        salt = os.urandom(16)
-    # Use PBKDF2 with SHA‑256 and 100_000 iterations (reasonable trade‑off)
-    hashed = hashlib.pbkdf2_hmac('sha256', password.encode('utf-8'), salt, 100_000)
-    return salt, hashed
-
-def verify_password(password: str, salt: bytes, expected_hash: bytes) -> bool:
-    """Verify a password against a stored salt and hash."""
-    _, hashed = hash_password(password, salt)
-    # Constant‑time comparison to avoid timing attacks
-    return hmac.compare_digest(hashed, expected_hash)
-
-
-def generate_invitation_code(length: int = 6) -> str:
-    alphabet = string.ascii_uppercase + string.digits
-    return ''.join(secrets.choice(alphabet) for _ in range(length))
-
-
-def generate_unique_invitation_code() -> str:
-    with get_db_connection() as conn:
-        cur = conn.cursor()
-        while True:
-            code = generate_invitation_code()
-            execute_write(cur, 'SELECT 1 FROM groups WHERE invitation_code = ?', (code,))
-            if cur.fetchone() is None:
-                return code
-
-def generate_session(user_id: int, group_id: int | None, duration_seconds: int = 7 * 24 * 3600) -> str:
-    """Create a new session token for a user and store it with the active
-    group in the database.  ``duration_seconds`` controls the cookie's
-    lifetime; default is one week."""
-    token = secrets.token_hex(32)
-    expires_at = int(time.time()) + duration_seconds
-    with get_db_connection() as conn:
-        cur = conn.cursor()
-        execute_write(cur, 
-            'INSERT INTO sessions (token, user_id, group_id, expires_at) VALUES (?, ?, ?, ?)',
-            (token, user_id, group_id, expires_at)
-        )
-        safe_commit(conn)
-    return token
-
-def get_user_by_session(token: str) -> dict | None:
-    """Retrieve the user associated with a session token.  Returns a dict
-    representing the user row or ``None`` if the session is invalid or
-    expired.  Expired sessions are removed from the database."""
-    if not token:
-        return None
-    with get_db_connection() as conn:
-        cur = conn.cursor()
-        # Remove expired sessions
-        execute_write(cur, 'DELETE FROM sessions WHERE expires_at <= ?', (int(time.time()),))
-        # Fetch the session
-        execute_write(cur, 
-            'SELECT user_id, group_id FROM sessions WHERE token = ?',
-            (token,)
-        )
-        row = cur.fetchone()
-        if not row:
-            safe_commit(conn)
-            return None
-        user_id = row['user_id']
-        group_id = row['group_id']
-        # Extend session expiry on each use (sliding window)
-        new_expires = int(time.time()) + 7 * 24 * 3600
-        execute_write(cur, 
-            'UPDATE sessions SET expires_at = ? WHERE token = ?',
-            (new_expires, token)
-        )
-        # Fetch the user along with their role
-        execute_write(cur, 
-            'SELECT id, username, role FROM users WHERE id = ?',
-            (user_id,)
-        )
-        user_row = cur.fetchone()
-        safe_commit(conn)
-    if user_row:
-        return {
-            'id': user_row['id'],
-            'username': user_row['username'],
-            'role': user_row['role'],
-            'group_id': group_id,
-        }
-    return None
-
-def delete_session(token: str) -> None:
-    """Invalidate a session by removing it from the database."""
-    with get_db_connection() as conn:
-        cur = conn.cursor()
-        execute_write(cur, 'DELETE FROM sessions WHERE token = ?', (token,))
-        safe_commit(conn)
-
-def log_event(user_id: int | None, action: str, metadata: dict | None = None) -> None:
-    """Insert an entry into the logs table."""
-    with get_db_connection() as conn:
-        cur = conn.cursor()
-        execute_write(cur, 
-            'INSERT INTO logs (user_id, action, metadata) VALUES (?, ?, ?)',
-            (user_id, action, json.dumps(metadata or {}))
-        )
-        safe_commit(conn)
-
-
-def sanitize_name(name: str) -> str | None:
-    """Return a sanitized file/display name or ``None`` if invalid."""
-    name = (name or '').strip()
-    if '..' in name or '/' in name or '\\' in name:
-        return None
-    return name
-
-
-def scan_for_viruses(file_bytes: bytes) -> bool:
-    """Scan ``file_bytes`` using an external command if configured.
-
-    The command is provided via the ``AV_SCAN_CMD`` environment variable and
-    should return a zero exit status when the file is clean.  If the variable
-    is unset, the file is assumed to be safe."""
-    cmd = os.environ.get('AV_SCAN_CMD')
-    if not cmd:
-        return True
-    try:
-        result = subprocess.run(shlex.split(cmd), input=file_bytes, capture_output=True)
-        return result.returncode == 0
-    except Exception:
-        return False
-
-
-def parse_multipart_form_data(data: bytes, content_type: str) -> tuple[dict, dict]:
-    """Parse ``multipart/form-data`` content from ``data``.
-
-    Returns a tuple ``(fields, files)`` where ``fields`` is a mapping of
-    field names to text values and ``files`` maps field names to dictionaries
-    containing ``filename``, ``content_type`` and ``content`` bytes.
-    ``content_type`` must be the value of the ``Content-Type`` header from the
-    original request.
-    """
-
-    parser = BytesParser(policy=default)
-    try:
-        message = parser.parsebytes(b"Content-Type: " + content_type.encode() + b"\r\n\r\n" + data)
-    except Exception:
-        return {}, {}
-
-    fields: dict[str, str] = {}
-    files: dict[str, dict] = {}
-    for part in message.iter_parts():
-        if part.get_content_disposition() != "form-data":
-            continue
-        name = part.get_param("name", header="content-disposition")
-        if not name:
-            continue
-        filename = part.get_filename()
-        payload = part.get_payload(decode=True)
-        if filename is not None:
-            files[name] = {
-                "filename": filename,
-                "content_type": part.get_content_type(),
-                "content": payload or b"",
-            }
-        else:
-            charset = part.get_content_charset() or "utf-8"
-            fields[name] = (payload or b"").decode(charset, errors="replace")
-    return fields, files
 
 
 def parse_audio_notes_json(data: str | None) -> dict:
@@ -3745,12 +3037,7 @@ class BandTrackHandler(BaseHTTPRequestHandler):
 #############################
 
 def run_server(host: str = '0.0.0.0', port: int = 8080):
-    migrate_to_multigroup()
     init_db()
-    migrate_performance_location()
-    migrate_suggestion_votes()
-    if _using_postgres():
-        migrate_sessions_group_id()
     ws_port = int(os.environ.get('WS_PORT', port + 1))
     threading.Thread(target=start_ws_server, args=(host, ws_port), daemon=True).start()
     server = ThreadingHTTPServer((host, port), BandTrackHandler)
