+++ tests/test_partitions.py
@@ -4,7 +4,7 @@ import threading
 import http.client
 import time
 import json
-import server
+import bandtrack.api as server
 
 
 def start_test_server(tmp_db_path=None):
