+++ tests/test_verify_password.py
@@ -2,7 +2,7 @@ import os
 import sys
 
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
-from server import hash_password, verify_password
+from bandtrack.auth import hash_password, verify_password
 
 
 def test_verify_password_valid_and_invalid():
